+++ src/test/java/roomescape/controller/dto/MemberReservationRequestTest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.reservation;
+package roomescape.controller.dto;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,7 +12,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.controller.dto.MemberReservationRequest;
 
 class MemberReservationRequestTest {
