+++ src/test/java/roomescape/global/CookieUtilsTest.java
@@ -1,4 +1,4 @@
-package roomescape.utils;
+package roomescape.global;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
-import roomescape.global.CookieUtils;
 
 class CookieUtilsTest {
