+++ src/test/java/roomescape/service/dto/theme/ThemeRequestTest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto.time;
+package roomescape.service.dto.theme;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
-import roomescape.service.dto.theme.ThemeRequest;
 
 class ThemeRequestTest {
